Values for cards collected, bug fixes,  events added
Created a function that gets the values of the selected cards and stores that data in a seperate array.

Fixed bug where the last card selected was not being displayed in the middle of the first 2 cards.

Re-named some buttons and values

Began programming what happens at the end of a hand